requirements.txt
certifi==2019.9.11
chardet==3.0.4
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-braces==1.13.0
django-filter==2.2.0
django-filters==0.2.1
django-jet==1.0.8
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
drf-generators==0.4.0
filters==1.3.2
google-api-python-client==1.4.1
httplib2==0.14.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==6.2.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
gunicorn==19.9.0
whitenoise==4.1.2
dj-database-url==0.5.0
Psycopg2==2.8.4

